:zap: IMPROVE: adapter runtime write to disk
