+++ packages/houdini-svelte/src/plugin/codegen/index.ts
@@ -1,7 +1,6 @@
 import { GenerateHookInput, fs, Config } from 'houdini'
 
 import { stores_directory, type_route_dir } from '../kit'
-import adapter from './adapter'
 import components from './components'
 import kit from './routes'
 import stores from './stores'
@@ -15,7 +14,6 @@ export default async function (input: PluginGenerateInput) {
 
 	// generate the files
 	await Promise.all([
-		adapter(input),
 		kit(input.framework, input),
 		stores(input),
 		components(input.framework, input),
